+++ src/pine/db/connections.clj
@@ -30,4 +30,3 @@
     (swap! connections assoc id connection)
     (prn connections)
     id))
-