+++ src/pine/db/main.clj
@@ -5,9 +5,13 @@
 (def connection-id "Currently selected connection" (atom nil))
 (def references "References indexed by the connection id" (atom {}))
 
+;; Memoization
+;;
 (def memoize-references? true)
 ;; (def memoize-references? false)
 
+;; Schema / Initialization
+;;
 (defn init-references
   "Get the references for a given key"
   [id]
@@ -18,5 +22,18 @@
      (swap! references assoc id (postgres/get-indexed-references id))
      (@references id))))
 
+;; Connections
+;;
+(defn set-connection [id]
+  (reset! connection-id id)
+  (init-references id)
+  id)
+
+;; Query
+;;
 (defn run-query [id query]
   (postgres/run-query id query))
+
+(defn get-connection-count [id]
+  (let [result (run-query id {:query "SELECT COUNT(*) as connection_count FROM pg_stat_activity" :params []})]
+    (-> result second first)))